requirements.txt
flet==0.24.1
python-dotenv>=1.0.0
google-genai>=0.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
python-pptx>=0.6.0
odfpy>=1.4.0
reportlab>=4.0.0
PyMuPDF>=1.23.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
